+++ connector/src/account_fetchers.rs
@@ -6,23 +6,62 @@ use std::sync::Mutex;
 
 use async_once_cell::unpin::Lazy;
 
-use anyhow::{Context, Error};
+use anyhow::{anyhow, Context, Error};
 
 
 use solana_client::nonblocking::rpc_client::RpcClient as RpcClientAsync;
 use solana_sdk::account::{AccountSharedData, ReadableAccount};
 use solana_sdk::pubkey::Pubkey;
+use crate::account_fetcher::AccountFetcherFeeds;
 
 
 #[async_trait::async_trait]
-pub trait AccountFetcherFeeds: Sync + Send {
-    async fn feeds_fetch_raw_account(&self, address: &Pubkey) -> anyhow::Result<AccountSharedData>;
+impl AccountFetcherFeeds for RpcAccountFetcher {
+    async fn feeds_fetch_raw_account(&self, address: &Pubkey) -> anyhow::Result<AccountSharedData> {
+        let sdfs: Result<AccountSharedData, anyhow::Error> = self.rpc
+            .get_account_with_commitment(address, self.rpc.commitment())
+            .await
+            .with_context(|| format!("fetch account {}", *address))?
+            .value
+            // .ok_or(ClientError::AccountNotFound)
+            .ok_or(anyhow!("Account not found")) // TODO is this correct?
+            .with_context(|| format!("fetch account {}", *address))
+            .map(Into::into);
+        sdfs
+    }
+
     async fn feeds_fetch_program_accounts(
         &self,
         program: &Pubkey,
         discriminator: [u8; 8],
-    ) -> anyhow::Result<Vec<(Pubkey, AccountSharedData)>>;
+    ) -> anyhow::Result<Vec<(Pubkey, AccountSharedData)>> {
+        use solana_account_decoder::UiAccountEncoding;
+        use solana_client::rpc_config::{RpcAccountInfoConfig, RpcProgramAccountsConfig};
+        use solana_client::rpc_filter::{Memcmp, RpcFilterType};
+        let config = RpcProgramAccountsConfig {
+            filters: Some(vec![RpcFilterType::Memcmp(Memcmp::new_raw_bytes(
+                0,
+                discriminator.to_vec(),
+            ))]),
+            account_config: RpcAccountInfoConfig {
+                encoding: Some(UiAccountEncoding::Base64),
+                commitment: Some(self.rpc.commitment()),
+                ..RpcAccountInfoConfig::default()
+            },
+            with_context: Some(true),
+        };
+        let accs = self
+            .rpc
+            .get_program_accounts_with_config(program, config)
+            .await?;
+        // convert Account -> AccountSharedData
+        Ok(accs
+            .into_iter()
+            .map(|(pk, acc)| (pk, acc.into()))
+            .collect::<Vec<_>>())
+    }
 }
+
 
 
 pub struct RpcAccountFetcher {
